Revert "linux/x11: Reduce input latency and ensure rerender priority (#13355)"
This reverts commit f69c8ca74eb934f37d29cbe2f6ee31f94c7aaf63.
