+++ crates/gpui/src/platform/linux/x11/window.rs
@@ -14,7 +14,6 @@ use util::{maybe, ResultExt};
 use x11rb::{
     connection::Connection,
     protocol::{
-        randr::{self, ConnectionExt as _},
         xinput::{self, ConnectionExt as _},
         xproto::{
             self, ClientMessageEvent, ConnectionExt as _, EventMask, TranslateCoordinatesReply,
@@ -32,7 +31,6 @@ use std::{
     ptr::NonNull,
     rc::Rc,
     sync::{self, Arc},
-    time::{Duration, Instant},
 };
 
 use super::{X11Display, XINPUT_MASTER_DEVICE};
@@ -161,8 +159,6 @@ pub struct Callbacks {
 
 pub struct X11WindowState {
     pub destroyed: bool,
-    pub last_render_at: Option<Instant>,
-    pub refresh_rate: Duration,
     client: X11ClientStatePtr,
     executor: ForegroundExecutor,
     atoms: XcbAtoms,
@@ -401,31 +397,6 @@ impl X11WindowState {
         };
         xcb_connection.map_window(x_window).unwrap();
 
-        let screen_resources = xcb_connection
-            .randr_get_screen_resources(x_window)
-            .unwrap()
-            .reply()
-            .expect("Could not find available screens");
-
-        let mode = screen_resources
-            .crtcs
-            .iter()
-            .find_map(|crtc| {
-                let crtc_info = xcb_connection
-                    .randr_get_crtc_info(*crtc, x11rb::CURRENT_TIME)
-                    .ok()?
-                    .reply()
-                    .ok()?;
-
-                screen_resources
-                    .modes
-                    .iter()
-                    .find(|m| m.id == crtc_info.mode)
-            })
-            .expect("Unable to find screen refresh rate");
-
-        let refresh_rate = mode_refresh_rate(mode);
-
         Ok(Self {
             client,
             executor,
@@ -442,8 +413,6 @@ impl X11WindowState {
             appearance,
             handle,
             destroyed: false,
-            last_render_at: None,
-            refresh_rate,
         })
     }
 
@@ -613,11 +582,6 @@ impl X11WindowStatePtr {
         let mut cb = self.callbacks.borrow_mut();
         if let Some(ref mut fun) = cb.request_frame {
             fun();
-
-            self.state
-                .borrow_mut()
-                .last_render_at
-                .replace(Instant::now());
         }
     }
 
@@ -1063,13 +1027,4 @@ impl PlatformWindow for X11Window {
     fn should_render_window_controls(&self) -> bool {
         false
     }
-}
-
-// Adatpted from:
-// https://docs.rs/winit/0.29.11/src/winit/platform_impl/linux/x11/monitor.rs.html#103-111
-pub fn mode_refresh_rate(mode: &randr::ModeInfo) -> Duration {
-    let millihertz = mode.dot_clock as u64 * 1_000 / (mode.htotal as u64 * mode.vtotal as u64);
-    let micros = 1_000_000_000 / millihertz;
-    log::info!("Refreshing at {} micros", micros);
-    Duration::from_micros(micros)
 }
